refactor: Merged component; extracted default data
