--- src/app/project/ontology/add-source-type/add-source-type.component.scss
@@ -1,7 +1,0 @@
-.form-content {
-  display: contents;
-
-  .select-source-type {
-    margin: 0 24px 0 0;
-  }
-}
