+++ src/app/project/ontology/ontology.component.html
@@ -43,7 +43,7 @@
 
                 <span class="app-toolbar-action" [class.select-form]="ontologies.length">
                     <button mat-raised-button color="primary" class="create-data-model-btn"
-                        (click)="openDialog('createOntology')">
+                        (click)="openOntologyForm('createOntology')">
                         {{ontologies.length ? 'New data model' : 'Create your first data model'}}
                     </button>
                 </span>
@@ -65,17 +65,30 @@
                 <span class="fill-remaining-space"></span>
 
                 <button mat-stroked-button
-                    (click)="$event.stopPropagation(); openDialog('editOntologyInfo', ontology.label, ontology.id)">
+                    (click)="$event.stopPropagation(); openOntologyForm('editOntologyInfo', ontology.label, ontology.id)">
                     Edit info
                 </button>
             </mat-toolbar-row>
         </mat-toolbar>
 
         <div #ontologyEditor class="ontology-editor-canvas" *ngIf="!loadOntology">
-            <!-- button to define and add new source type to the ontology -->
-            <app-add-source-type *ngIf="project?.status && sysAdmin && ontologyIri" [projectcode]="projectcode"
+            <!-- button to select and add new source type to the ontology -->
+            <div class="add-source-type" *ngIf="project?.status && sysAdmin && ontologyIri">
+                <button mat-raised-button color="primary" [matMenuTriggerFor]="beforeMenu">
+                    Add source type&nbsp;
+                    <mat-icon>add</mat-icon>
+                </button>
+                <mat-menu #beforeMenu="matMenu" xPosition="before">
+                    <button mat-menu-item *ngFor="let type of sourceTypes"
+                        (click)="openSourceTypeForm('addSourceType', type)">
+                        {{ type?.label }}
+                    </button>
+                </mat-menu>
+            </div>
+
+            <!-- <app-add-source-type *ngIf="project?.status && sysAdmin && ontologyIri" [projectcode]="projectcode"
                 class="add-source-type" (refreshParent)="resetOntology(ontologyIri)" #addSourceTypeComponent>
-            </app-add-source-type>
+            </app-add-source-type> -->
 
             <!-- list of source types -->
             <!-- <svg width="100%" height="1200" class="mySvg" (click)="clicked($event)"></svg> -->
@@ -104,58 +117,6 @@
     <app-error [status]="403"></app-error>
 </div>
 
-
-<!--
-
-<div class="content ontology-editor drag-drop-boundary desktop-only">
-
-    <mat-card class="resource-type" cdkDragBoundary=".drag-drop-boundary" cdkDrag>
-        Create your data model: add the first source type here
-    </mat-card>
-
-
-    <!--
-        <div id="ontology-editor" class="ontology-editor" >
-                Create your data model first:
-                - Drag'n'drop your source types from the list on the right hand side.
-                - You can customize the predefined sources to your needs later.
-        </div>
-    -->
-<!--
-    <mat-toolbar>
-                <span>Select ontology to edit</span>
-                <span class="fill-remaining-space"></span>
-                <span>
-                    <mat-icon>add</mat-icon>
-                    Add new source type</span>
-            </mat-toolbar>
-
-
-    <div id="ontology-editor" class="ontology-editor" >
-        Create your data model first:
-        - Drag'n'drop your source types from the list on the right hand side.
-        - You can customize the predefined sources to your needs later.
-        <mat-accordion #ontologyEditor>
-            <app-resource-type></app-resource-type>
-        </mat-accordion>
-
-    </div>
-
-    <div class="ontology-default-elements">
-        Select your source types from this list:<br><br>
-
-        <div class="default-source-types">
-            <div class="list-item"
-              *ngFor="let st of sourcetypes"
-              (click)="loadComponent()">
-              <mat-icon>keyboard_arrow_left</mat-icon>
-                  {{st}}
-            </div>
-          </div>
-
-    </div>
-</div>
--->
 
 <div class="mobile-only">
     <kui-message [message]="{status: 415, statusText: 'This content is not supported on small devices.'}">
