+++ src/app/project/ontology/source-type-form/source-type-form.component.html
@@ -5,16 +5,19 @@
     <!-- success message after updating -->
     <kui-message *ngIf="success" [message]="successMessage" [short]="true" (click)="closeMessage()"></kui-message>
 
-    <!-- content -->
-    <form *ngIf="!loading" [formGroup]="sourceTypeForm" class="form-content project-data" (ngSubmit)="submitData()">
+    <!-- content: Step 1 = define resource type -->
+    <form *ngIf="!loading && showSourceTypeForm" [formGroup]="sourceTypeForm" class="form-content source-type-data"
+        (ngSubmit)="submitSourceType()">
 
         <!-- label -->
-        <mat-form-field class="large-field hidden">
+        <!-- <mat-form-field class="large-field hidden">
             <input matInput placeholder="Subclass of *" formControlName="subClassOf" [value]="subClassOf" readonly>
-        </mat-form-field>
+        </mat-form-field> -->
 
         <div class="medium-field string-literal-container">
-            <kui-string-literal-input [placeholder]="'Label *'"></kui-string-literal-input>
+            <kui-string-literal-input [placeholder]="'Label *'" [value]="sourceTypeLabels"
+                (dataChanged)="handleData($event, 'labels')">
+            </kui-string-literal-input>
             <mat-hint class="string-literal-error" *ngIf="formErrors.label">
                 {{ formErrors.label }}
             </mat-hint>
@@ -40,7 +43,8 @@
         <br>
         <!-- description -->
         <div class="large-field string-literal-container">
-            <kui-string-literal-input [placeholder]="'Description'" [textarea]="true"></kui-string-literal-input>
+            <kui-string-literal-input [placeholder]="'Description'" [value]="sourceTypeComments" [textarea]="true"
+                (dataChanged)="handleData($event, 'comments')"></kui-string-literal-input>
         </div>
 
         <!-- <mat-form-field class="large-field">
@@ -50,19 +54,58 @@
         <!-- action buttons -->
         <div class="form-panel form-action">
             <span>
-                <button mat-button type="reset" (click)="resetForm($event)">
+                <button mat-button type="reset" (click)="resetSourceTypeForm($event)">
                     {{ 'appLabels.form.action.reset' | translate }}
                 </button>
             </span>
             <span class="fill-remaining-space"></span>
             <span>
-                <button mat-raised-button type="submit" color="primary" [disabled]="!formValid">
+                <button mat-raised-button type="submit" color="primary" [disabled]="!sourceTypeLabels.length">
                     Next
                 </button>
             </span>
         </div>
     </form>
+
+    <!-- content: Step 2 = define properties for resource type -->
+    <form *ngIf="!loading && !showSourceTypeForm" [formGroup]="sourceTypeFormSub" class="form-content properties-form"
+        (ngSubmit)="submitProperties()">
+
+        <h3 class="mat-body subtitle">Define the metadata for this source type</h3>
+
+        <!-- TODO: properties -->
+        <!-- one line, with "add" icon to add additional properties -->
+        <div cdkDropList class="properties-list" (cdkDropListDropped)="drop($event)">
+            <div cdkDrag class="property-line"
+                *ngFor="let prop of properties?.controls; let last = last; let i = index;">
+
+                <div class="drag-n-drop-placeholder" *cdkDragPlaceholder></div>
+
+                <button *ngIf="properties.length > 1" cdkDragHandle mat-icon-button type="button"
+                    aria-label="Change order of properties" class="prefix drag-n-drop-line">
+                    <mat-icon>drag_indicator</mat-icon>
+                </button>
+
+                <app-source-type-property [propertyForm]="prop" [index]="i">
+                </app-source-type-property>
+
+                <button *ngIf="properties.length > 1" mat-icon-button type="button" color="warn"
+                    aria-label="Delete this property" class="suffix delete-line" (click)="removeProperty(i)">
+                    <mat-icon>delete_outlined</mat-icon>
+                </button>
+                <button *ngIf="last" [disabled]="!properties.valid" mat-icon-button type="button" color="primary"
+                    aria-label="Add new property" class="add-new-line" (click)="addProperty()">
+                    <mat-icon>add</mat-icon>
+                </button>
+            </div>
+
+        </div>
+    </form>
+
+
 </div>
+
+
 
 
 <!-- in the case of an API error: show the error message -->
